added todo to refactor TestCaseInfoTests
