+++ src/test/java/cloud/autotests/tests/testcase/HistoryTests.java
@@ -1,0 +1,4 @@
+package cloud.autotests.tests.testcase;
+
+public class HistoryTests {
+}
