+++ src/test/java/cloud/autotests/tests/testcase/TestCasesInfoTests.java
@@ -1,4 +1,4 @@
-package cloud.autotests.tests;
+package cloud.autotests.tests.testcase;
 
 import cloud.autotests.config.App;
 import cloud.autotests.data.MenuItem;
@@ -8,9 +8,11 @@ import cloud.autotests.pages.ProjectPage;
 import cloud.autotests.pages.ProjectsTable;
 import cloud.autotests.pages.TestCasesTable;
 import cloud.autotests.pages.components.Sidebar;
+import cloud.autotests.tests.TestBase;
 import com.codeborne.selenide.Condition;
 import com.codeborne.selenide.SelenideElement;
 import io.qameta.allure.Step;
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
 import static com.codeborne.selenide.CollectionCondition.size;
@@ -21,12 +23,14 @@ import static com.codeborne.selenide.Selenide.*;
 import static io.qameta.allure.Allure.step;
 import static org.assertj.core.api.Assertions.assertThat;
 
+
 public class TestCasesInfoTests extends TestBase {
     private static final String PROJECT_NAME = "teacher qa_guru_diplom_project";
 
 
     @WithLogin
     @Test
+    @Disabled("Cut to different test classes")
     void TestCasesPageShouldContainValidInfo() {
         ProjectsTable projectsTable = open(App.config.webUrl(), ProjectsTable.class);
         ProjectPage projectPage = projectsTable.navigateTo(PROJECT_NAME);
@@ -36,14 +40,6 @@ public class TestCasesInfoTests extends TestBase {
         step("Open active test case", () -> {
             $$(".LoadableTree__view > li").findBy(text("Active")).click();
             $(".TestCaseLayout__name").shouldBe(visible);
-        });
-
-        step("Add and delete comment", () -> {
-            $(".MarkdownTextarea__edit").setValue("test comment");
-            $(".Button_style_success ").click();
-
-            $(".Comment__header").$("svg").click();
-            $(".tippy-content").find(byText("Delete")).click();
         });
 
         step("Open History", () -> {
