+++ src/test/resources/config/local.properties
@@ -3,5 +3,6 @@ browserVersion=91.0
 browserSize=1920x1080
 browserMobileView=
 #browserMobileView=iPhone X
-remoteDriverUrl=https://user1:1234@selenoid.autotests.cloud/wd/hub/
+remoteDriverUrl=
+#remoteDriverUrl=https://user1:1234@selenoid.autotests.cloud/wd/hub/
 videoStorage=
